Adding callback to create paragraph after_save Writing
Paragraphs are created, including writing_id, par_num (paragraph number) and body, which includes the html tags we will have to eliminate int he view.

We will need to refactor the view/controller to use the model methods instead of the gem's to return the paragraphs for the view. which seems better.

for Issue #9