Allowing omments to be created via form
User can now create a comment by clicking on the comment icon.

When the comment is created, the user is redirected to the comments index page, which shows all the comments related to that specific paragraph.  Arguably, this should redirect instead to the paragraph show view, so we need to figure that out.  In the meantime, I have simply adjusted the comments index view to show the content. this is not intended to be a permanent solution.

In addition to fixing the redirect, I want to associate a user_id with a comment, such that every comment belogs to a user.